add overlap to integral
